refactor(llm): repoint production consumers to the published package root (#1426 — commit 2/8)
All src/ consumers now import via 'from src.llm import …' (the published
interface); deep layer paths remain only inside the package and in tests
that exercise internals.
