+++ apps/backend/src/services/ai_streaming.py
@@ -15,12 +15,22 @@ from typing import Any
 from uuid import UUID
 
 from src.config import settings
-from src.llm import protocol_for
-from src.llm.base import LLMConfigError, LLMError, ProtocolFamily, ProviderRef, ReasoningEffort
-from src.llm.base.usage import estimate_tokens, estimate_tokens_from_chars
-from src.llm.extension.cassette import CassetteMode, current_mode
-from src.llm.extension.client import litellm_stream, resolve_provider_and_model
-from src.llm.extension.factory import get_config_source, get_usage_meter
+from src.llm import (
+    CassetteMode,
+    LLMConfigError,
+    LLMError,
+    ProtocolFamily,
+    ProviderRef,
+    ReasoningEffort,
+    current_mode,
+    estimate_tokens,
+    estimate_tokens_from_chars,
+    get_config_source,
+    get_usage_meter,
+    litellm_stream,
+    protocol_for,
+    resolve_provider_and_model,
+)
 from src.observability import get_logger, record_ai_provider_call
 
 logger = get_logger(__name__)
